refactor: move AST node builder functions to separate internal-only module
